fix(tir): change /v4/issuers implementation to match specification
The specification indicates that page[after] represents the page being accessed. The implementation has been updated to follow this behavior. Pages start at 0.
Additionally, the links and the self did not reference the full URL, but only the path, without the host, port, or protocol. For its implementation,
`micronaut.server.forward-headers` configuration has been added to read the X-Forwarded-* headers.